.github/workflows/publish.yml: Run after `gentoo/stage3` update time
The gentoo/stage3 image is set to be built and updated in GitHub Actions
at 1:00 every day, but this image has been built at 22:00, which means
each build would be outdated in just about 3 hours.

The new build time of this image is 2:30 because although the
gentoo/stage3 is scheduled at 1:00, it might not be picked up by a
GitHub Actions runner until around 1:50.

Signed-off-by: Leo <liaoyuan@gmail.com>