Add JDK 11 to the image
When some dependencies of JDK 11 are being installed in a container, the
container needs to have some extra privileges granted to it, or else the
dependencies would fail to install.  Because Docker does not support
granting extra privileges to a container created and used to build an
image, it can no longer be used to build this image once JDK 11 has been
added to it; Buildah, which supports this in contrast, is now being used
instead.

Signed-off-by: Yuan Liao <liaoyuan@gmail.com>